Refactor: Modularize Lab Page - Extract UploadModal and related logic
- Extract types to lib/lab/types.ts
- Extract file utilities to lib/lab/fileUtils.ts
- Extract upload utilities to lib/lab/uploadUtils.ts
- Create useDragAndDrop hook for drag/drop functionality
- Create usePasteHandler hook for paste event handling
- Create useFileUpload hook for file management
- Create useUploadForm hook for form state and submission
- Extract UploadModal component to components/lab/UploadModal.tsx
- Refactor main page.tsx to use extracted components and hooks

Results:
- Reduced page.tsx from 1141 lines to 328 lines (71% reduction)
- Improved maintainability with single-responsibility modules
- Enhanced reusability of hooks and utilities
- Better testability with separated logic
- No behavior changes - all functionality preserved
