+++ app/lab/[id]/page.tsx
@@ -4,9 +4,7 @@ import React, { useEffect, useState } from "react";
 import { useRouter, useParams } from 'next/navigation';
 import Link from 'next/link';
 import Navigation from "@/components/navigation";
-import { detectLanguage } from "@/lib/language-detection";
-import { Switch } from "@/components/ui/switch";
-import { Label } from "@/components/ui/label";
+import { UploadModal } from "@/components/lab/UploadModal";
 import {
   Breadcrumb,
   BreadcrumbList,
@@ -15,44 +13,7 @@ import {
   BreadcrumbPage,
   BreadcrumbSeparator,
 } from "@/components/ui/breadcrumb";
-
-interface Course {
-  id: string;
-  name: string;
-  description: string;
-}
-
-interface Lab {
-  id: string;
-  course_id: string;
-  lab_number: number;
-  title: string;
-  description: string;
-  courses?: Course;
-}
-
-interface Student {
-  id: string;
-  name: string;
-  email: string;
-  track_id: string;
-}
-
-interface Track {
-  id: string;
-  code: string;
-  name: string;
-}
-
-interface Submission {
-  id: string;
-  title: string;
-  student_id: string;
-  view_count: number;
-  upvote_count?: number;
-  is_anonymous?: boolean;
-  students?: Student;
-}
+import type { Lab, Student, Track, Submission } from "@/lib/lab/types";
 
 export default function LabPage() {
   const [lab, setLab] = useState<Lab | null>(null);
@@ -362,779 +323,5 @@ export default function LabPage() {
         </div>
       </div>
     </div>
-  );
-}
-
-interface UploadModalProps {
-  labId: string;
-  onClose: (shouldNavigateBack?: boolean) => void;
-}
-
-function UploadModal({ labId, onClose }: UploadModalProps) {
-  // Array of random solution name variations
-  const randomSolutionNames = [
-    "Solution",
-    "Approach",
-    "Submission",
-    "Lab",
-    "Task",
-    "Assignment",
-    "Implementation",
-    "Method",
-    "Answer",
-    "Response",
-    "Attempt",
-    "Draft",
-    "Work",
-    "Result",
-    "Entry",
-    "Try",
-    "Effort",
-    "Take",
-    "Way",
-    "Path",
-    "Plan",
-    "Idea",
-    "Fix",
-    "Reply",
-    "Turn",
-    "Shot",
-    "Go",
-    "Run",
-    "Pass",
-    "Hit",
-    "Crack",
-    "Break",
-    "Win",
-    "Nail",
-    "Smash",
-    "Crush",
-    "Rock",
-    "Boss",
-    "Beast",
-    "Fire",
-    "Wave",
-    "Vibe",
-    "Move",
-    "Flex",
-    "Show"
-  ];
-
-  // Get a random solution name on mount
-  const getRandomSolutionName = () => {
-    return randomSolutionNames[Math.floor(Math.random() * randomSolutionNames.length)];
-  };
-
-  const [title, setTitle] = useState(getRandomSolutionName());
-  const [pastedContent, setPastedContent] = useState("");
-  const [pastedFileName, setPastedFileName] = useState("");
-  const [language, setLanguage] = useState("text");
-  const [files, setFiles] = useState<File[]>([]);
-  const [pastedCodeFiles, setPastedCodeFiles] = useState<Array<{ filename: string; language: string; content: string }>>([]);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState("");
-  const [isDragging, setIsDragging] = useState(false);
-  const [isDraggingOverPage, setIsDraggingOverPage] = useState(false);
-  const [showPasteArea, setShowPasteArea] = useState(false);
-  const [isAnonymous, setIsAnonymous] = useState(false);
-  const fileInputRef = React.useRef<HTMLInputElement>(null);
-  const pasteTextareaRef = React.useRef<HTMLTextAreaElement>(null);
-  const titleInputRef = React.useRef<HTMLInputElement>(null);
-  const dragCounterRef = React.useRef(0);
-
-  const detectedLanguages: Record<string, string> = {
-    ".js": "javascript",
-    ".ts": "typescript",
-    ".py": "python",
-    ".cpp": "cpp",
-    ".c": "c",
-    ".java": "java",
-    ".cs": "csharp",
-    ".php": "php",
-    ".rb": "ruby",
-    ".go": "go",
-    ".rs": "rust",
-    ".txt": "text",
-    ".sql": "SQL",
-    ".html": "HTML",
-    ".css": "CSS",
-  };
-
-  const codeExtensions = ['.js', '.ts', '.py', '.cpp', '.c', '.java', '.cs', '.php', '.rb', '.go', '.rs', '.txt', '.sql', '.html', '.css'];
-
-  // Helper function to convert file to base64
-  const fileToBase64 = (file: File): Promise<string> => {
-    return new Promise((resolve, reject) => {
-      const reader = new FileReader();
-      reader.readAsDataURL(file);
-      reader.onload = () => resolve(reader.result as string);
-      reader.onerror = (error) => reject(error);
-    });
-  };
-
-  // Add files to the list (avoid duplicates)
-  const addFiles = (newFiles: File[]) => {
-    setFiles((prev) => {
-      // Check against both existing files and pasted code files
-      const existingFileNames = new Set(prev.map(f => f.name.toLowerCase()));
-      const existingPastedNames = new Set(pastedCodeFiles.map(f => f.filename.toLowerCase()));
-      const allExistingNames = new Set([...existingFileNames, ...existingPastedNames]);
-      
-      const uniqueNewFiles = newFiles.filter(f => !allExistingNames.has(f.name.toLowerCase()));
-      
-      if (uniqueNewFiles.length < newFiles.length) {
-        const duplicates = newFiles.filter(f => allExistingNames.has(f.name.toLowerCase()));
-        setError(`Skipped ${newFiles.length - uniqueNewFiles.length} duplicate file(s): ${duplicates.map(f => f.name).join(', ')}`);
-      } else {
-        setError("");
-      }
-      
-      return [...prev, ...uniqueNewFiles];
-    });
-  };
-
-  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    if (e.target.files) {
-      addFiles(Array.from(e.target.files));
-    }
-  };
-
-  // Global drag handlers for page-level drag detection
-  React.useEffect(() => {
-    const handleGlobalDragEnter = (e: DragEvent) => {
-      // Only show overlay if dragging files (not text/elements)
-      if (e.dataTransfer?.types.includes('Files')) {
-        dragCounterRef.current++;
-        if (dragCounterRef.current === 1) {
-          setIsDraggingOverPage(true);
-        }
-      }
-    };
-
-    const handleGlobalDragLeave = (e: DragEvent) => {
-      // Only decrement if we're actually leaving the document
-      // Check if we're leaving to a point outside the viewport
-      if (e.clientX === 0 && e.clientY === 0) {
-        dragCounterRef.current = 0;
-        setIsDraggingOverPage(false);
-      } else {
-        dragCounterRef.current--;
-        if (dragCounterRef.current <= 0) {
-          dragCounterRef.current = 0;
-          setIsDraggingOverPage(false);
-        }
-      }
-    };
-
-    const handleGlobalDragOver = (e: DragEvent) => {
-      // Only prevent default if dragging files
-      if (e.dataTransfer?.types.includes('Files')) {
-        e.preventDefault();
-      }
-    };
-
-    const handleGlobalDrop = (e: DragEvent) => {
-      // Reset counter when files are dropped anywhere
-      dragCounterRef.current = 0;
-      setIsDraggingOverPage(false);
-    };
-
-    // Add event listeners to document
-    document.addEventListener('dragenter', handleGlobalDragEnter);
-    document.addEventListener('dragleave', handleGlobalDragLeave);
-    document.addEventListener('dragover', handleGlobalDragOver);
-    document.addEventListener('drop', handleGlobalDrop);
-
-    return () => {
-      document.removeEventListener('dragenter', handleGlobalDragEnter);
-      document.removeEventListener('dragleave', handleGlobalDragLeave);
-      document.removeEventListener('dragover', handleGlobalDragOver);
-      document.removeEventListener('drop', handleGlobalDrop);
-    };
-  }, []);
-
-  // Drag and drop handlers for the drop zone
-  const handleDragEnter = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(true);
-  };
-
-  const handleDragLeave = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    // Only set dragging to false if we're actually leaving the drop zone
-    const rect = e.currentTarget.getBoundingClientRect();
-    const x = e.clientX;
-    const y = e.clientY;
-    
-    if (x < rect.left || x > rect.right || y < rect.top || y > rect.bottom) {
-      setIsDragging(false);
-    }
-  };
-
-  const handleDragOver = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-  };
-
-  const handleDrop = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(false);
-    setIsDraggingOverPage(false);
-    dragCounterRef.current = 0;
-
-    const droppedFiles = Array.from(e.dataTransfer.files);
-    if (droppedFiles.length > 0) {
-      addFiles(droppedFiles);
-    }
-  };
-
-  // Handle paste (Ctrl+V) for files/images and text
-  React.useEffect(() => {
-    const handlePaste = (e: ClipboardEvent) => {
-      const items = e.clipboardData?.items;
-      if (!items) return;
-
-      // Don't interfere if user is pasting into the textarea itself (let normal paste behavior happen)
-      const target = e.target as HTMLElement;
-      if (target && (target.tagName === 'TEXTAREA' || target.tagName === 'INPUT')) {
-        return;
-      }
-
-      // Check for images first (clipboard images)
-      const filesToAdd: File[] = [];
-      let hasText = false;
-      
-      for (let i = 0; i < items.length; i++) {
-        const item = items[i];
-        
-        // Check for image files (from clipboard)
-        if (item.kind === 'file' && item.type.startsWith('image/')) {
-          const file = item.getAsFile();
-          if (file) {
-            filesToAdd.push(file);
-          }
-        } else if (item.kind === 'string' && item.type === 'text/plain') {
-          hasText = true;
-        }
-      }
-
-      // If images found, add them as attachments
-      if (filesToAdd.length > 0) {
-        e.preventDefault();
-        addFiles(filesToAdd);
-      } else if (hasText) {
-        // If text is pasted, open text area and paste text directly
-        e.preventDefault();
-        
-        // Get text content synchronously from clipboard
-        const textContent = e.clipboardData?.getData('text/plain') || '';
-        
-        // Open paste area if not already open
-        if (!showPasteArea) {
-          setShowPasteArea(true);
-        }
-        
-        // Wait a bit for textarea to be available, then set the text
-        setTimeout(() => {
-          if (textContent) {
-            setPastedContent(textContent);
-            // Auto-detect language from pasted content
-            const detectedLang = detectLanguage(textContent);
-            setLanguage(detectedLang);
-            pasteTextareaRef.current?.focus();
-            // Move cursor to end
-            if (pasteTextareaRef.current) {
-              const textarea = pasteTextareaRef.current;
-              textarea.setSelectionRange(textarea.value.length, textarea.value.length);
-            }
-          }
-        }, 10);
-      }
-    };
-
-    window.addEventListener('paste', handlePaste);
-    return () => window.removeEventListener('paste', handlePaste);
-  }, [showPasteArea]);
-
-  // Handle text paste in textarea - let normal paste behavior happen
-  const handleTextareaPaste = (e: React.ClipboardEvent<HTMLTextAreaElement>) => {
-    // Normal paste behavior is allowed - React will handle it automatically
-    // We don't need to do anything special here
-  };
-
-  // Cancel paste area
-  const handleCancelPaste = () => {
-    setShowPasteArea(false);
-    setPastedContent("");
-    setPastedFileName("");
-  };
-
-  // Add pasted code as a file
-  const handleAddPastedCode = () => {
-    if (!pastedContent.trim()) {
-      setError("Please paste some code first");
-      return;
-    }
-    
-    const filename = pastedFileName.trim() || `code.${language}`;
-    
-    // Check for duplicate filenames (case-insensitive)
-    const existingFileNames = new Set(files.map(f => f.name.toLowerCase()));
-    const existingPastedNames = new Set(pastedCodeFiles.map(f => f.filename.toLowerCase()));
-    const allExistingNames = new Set([...existingFileNames, ...existingPastedNames]);
-    
-    if (allExistingNames.has(filename.toLowerCase())) {
-      setError(`A file with the name "${filename}" already exists. Please use a different name.`);
-      return;
-    }
-    
-    setPastedCodeFiles(prev => [...prev, {
-      filename,
-      language,
-      content: pastedContent,
-    }]);
-    
-    setPastedContent("");
-    setPastedFileName("");
-    setError("");
-  };
-
-  // Remove file from list
-  const removeFile = (index: number) => {
-    setFiles((prev) => prev.filter((_, i) => i !== index));
-  };
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setLoading(true);
-    setError("");
-
-    try {
-      // Validate title - if user deleted the default, require them to enter one
-      if (!title.trim() || title.trim() === "") {
-        setError("Solution name is required");
-        setLoading(false);
-        return;
-      }
-
-      let filesToUpload: Array<
-        | { filename: string; language: string; content: string }
-        | { filename: string; mimeType: string; base64: string }
-      > = [];
-
-      // Add pasted code files
-      filesToUpload.push(...pastedCodeFiles);
-
-      // Process uploaded/dropped files
-      if (files.length > 0) {
-        for (const file of files) {
-          const ext = "." + file.name.split(".").pop()?.toLowerCase();
-          
-          // Check if it's a code file
-          if (codeExtensions.includes(ext)) {
-            // Parse as code file
-            const content = await file.text();
-            const detectedLang = detectedLanguages[ext] || "text";
-
-            filesToUpload.push({
-              filename: file.name,
-              language: detectedLang,
-              content,
-            });
-          } else {
-            // Convert to base64 for attachment
-            const base64 = await fileToBase64(file);
-            filesToUpload.push({
-              filename: file.name,
-              mimeType: file.type || 'application/octet-stream',
-              base64: base64,
-            });
-          }
-        }
-      }
-
-      // Validate that we have at least one file
-      if (filesToUpload.length === 0) {
-        setError("Please add at least one file");
-        setLoading(false);
-        return;
-      }
-
-      // Check for duplicate filenames (case-insensitive)
-      const seenFilenames = new Set<string>();
-      const duplicates: string[] = [];
-      
-      for (const file of filesToUpload) {
-        const filename = file.filename.toLowerCase();
-        if (seenFilenames.has(filename)) {
-          duplicates.push(file.filename);
-        } else {
-          seenFilenames.add(filename);
-        }
-      }
-      
-      if (duplicates.length > 0) {
-        setError(`Duplicate filenames detected: ${duplicates.join(', ')}. Please remove duplicates before uploading.`);
-        setLoading(false);
-        return;
-      }
-
-      // Upload submission via API route (server-side validation)
-      // Note: studentId is not sent - it's validated from the cookie server-side
-      const response = await fetch("/api/submission/upload", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        credentials: "include",
-        body: JSON.stringify({
-          labId,
-          title,
-          files: filesToUpload,
-          isAnonymous: isAnonymous,
-        }),
-      });
-
-      if (!response.ok) {
-        const errorData = await response.json();
-        throw new Error(errorData.error || "Failed to upload submission");
-      }
-
-      onClose();
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "Upload failed");
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  // Remove pasted code file
-  const removePastedCodeFile = (index: number) => {
-    setPastedCodeFiles(prev => prev.filter((_, i) => i !== index));
-  };
-
-  // Handle backdrop click - only close if clicking directly on backdrop, not modal content
-  const handleBackdropClick = (e: React.MouseEvent<HTMLDivElement>) => {
-    // Only close if clicking directly on the backdrop (not on modal content)
-    // Don't close if user has text selected (they might be selecting text in an input)
-    const selection = window.getSelection();
-    const hasSelection = selection && selection.toString().length > 0;
-    
-    if (e.target === e.currentTarget && !hasSelection) {
-      onClose();
-    }
-  };
-
-  return (
-    <>
-      {/* Global drag overlay - appears when dragging files anywhere on the page */}
-      {isDraggingOverPage && (
-        <div 
-          className="fixed inset-0 bg-white/80 backdrop-blur-md z-[60] flex items-center justify-center pointer-events-none animate-fade-in"
-        >
-          <div className="bg-gradient-card border-4 border-primary border-dashed rounded-2xl p-12 max-w-2xl mx-4 shadow-modern-xl transform transition-all duration-300 scale-105 animate-scale-in backdrop-blur-sm">
-            <div className="text-center">
-              <svg 
-                className="mx-auto h-16 w-16 text-primary mb-4 animate-bounce" 
-                fill="none" 
-                stroke="currentColor" 
-                viewBox="0 0 24 24"
-              >
-                <path 
-                  strokeLinecap="round" 
-                  strokeLinejoin="round" 
-                  strokeWidth={2} 
-                  d="M7 16a4 4 0 01-.88-7.903A5 5 0 1115.9 6L16 6a5 5 0 011 9.9M15 13l-3-3m0 0l-3 3m3-3v12" 
-                />
-              </svg>
-              <h3 className="text-lg sm:text-2xl font-bold text-foreground mb-2">Drop files to upload</h3>
-              <p className="text-sm sm:text-base text-muted-foreground">Release to add files to your submission</p>
-            </div>
-          </div>
-        </div>
-      )}
-
-      <div 
-        className="fixed inset-0 bg-white/80 backdrop-blur-md flex items-center justify-center z-50 p-2 sm:p-4 overflow-y-auto animate-fade-in"
-        onClick={handleBackdropClick}
-      >
-        <div 
-          className="bg-gradient-card border border-border/50 p-5 sm:p-7 max-w-lg w-full my-4 max-h-[90vh] overflow-y-auto rounded-2xl shadow-modern-xl backdrop-blur-sm animate-scale-in"
-          onClick={(e) => e.stopPropagation()}
-        >
-        <h2 className="text-xl sm:text-2xl font-bold text-foreground mb-4 sm:mb-5 bg-gradient-to-r from-foreground to-foreground/70 bg-clip-text">Upload Solution</h2>
-
-        <form onSubmit={handleSubmit} onClick={(e) => e.stopPropagation()} className="space-y-3 sm:space-y-4">
-          {/* Title */}
-          <div>
-            <label className="block text-sm sm:text-base text-foreground font-semibold mb-2.5">Solution Name</label>
-            <input
-              ref={titleInputRef}
-              type="text"
-              value={title}
-              onChange={(e) => setTitle(e.target.value)}
-              onClick={(e) => e.stopPropagation()}
-              onFocus={(e) => e.stopPropagation()}
-              onMouseDown={(e) => e.stopPropagation()}
-              onMouseUp={(e) => e.stopPropagation()}
-              onMouseMove={(e) => e.stopPropagation()}
-              onSelect={(e) => e.stopPropagation()}
-              onDragStart={(e) => e.stopPropagation()}
-              className="w-full px-4 py-3 text-sm sm:text-base border border-border/50 bg-white/80 text-foreground rounded-lg focus:outline-none focus:ring-2 focus:ring-primary/50 focus:border-primary shadow-modern transition-all duration-300 backdrop-blur-sm hover:border-primary/30"
-              required
-            />
-          </div>
-
-          {/* Anonymous Toggle */}
-          <div className="flex items-center gap-2">
-            <Switch
-              id="submission-anonymous"
-              checked={isAnonymous}
-              onCheckedChange={setIsAnonymous}
-              disabled={loading}
-            />
-            <Label htmlFor="submission-anonymous" className="text-sm text-foreground cursor-pointer">
-              Submit anonymously
-            </Label>
-          </div>
-
-          {/* Combined Upload & Paste Box */}
-          <div>
-            <label className="block text-sm sm:text-base text-foreground font-semibold mb-2.5">Files & Code</label>
-            <div
-              onDragEnter={handleDragEnter}
-              onDragOver={handleDragOver}
-              onDragLeave={handleDragLeave}
-              onDrop={handleDrop}
-              className={`border-2 border-dashed rounded-xl p-5 sm:p-7 lg:p-9 transition-all duration-300 ${
-                isDragging 
-                  ? 'border-primary bg-primary/10 scale-105 shadow-primary-lg' 
-                  : 'border-border/50 bg-white/80 hover:bg-accent/30 hover:border-primary/40 shadow-modern backdrop-blur-sm'
-              }`}
-            >
-              <div className="text-center">
-                {isDragging ? (
-                  <>
-                    <svg 
-                      className="mx-auto h-12 w-12 text-primary mb-3 animate-pulse" 
-                      fill="none" 
-                      stroke="currentColor" 
-                      viewBox="0 0 24 24"
-                    >
-                      <path 
-                        strokeLinecap="round" 
-                        strokeLinejoin="round" 
-                        strokeWidth={2} 
-                        d="M7 16a4 4 0 01-.88-7.903A5 5 0 1115.9 6L16 6a5 5 0 011 9.9M15 13l-3-3m0 0l-3 3m3-3v12" 
-                      />
-                    </svg>
-                    <p className="text-lg font-semibold text-foreground mb-1">
-                      Drop files here
-                    </p>
-                    <p className="text-sm text-gray-600">
-                      Release to upload
-                    </p>
-                  </>
-                ) : (
-                  <>
-                    <svg 
-                      className="mx-auto h-10 w-10 text-gray-400 mb-3" 
-                      fill="none" 
-                      stroke="currentColor" 
-                      viewBox="0 0 24 24"
-                    >
-                      <path 
-                        strokeLinecap="round" 
-                        strokeLinejoin="round" 
-                        strokeWidth={2} 
-                        d="M7 16a4 4 0 01-.88-7.903A5 5 0 1115.9 6L16 6a5 5 0 011 9.9M15 13l-3-3m0 0l-3 3m3-3v12" 
-                      />
-                    </svg>
-                    <p className="text-sm text-foreground mb-2">
-                      <span className="font-semibold">Drag files here</span> or{' '}
-                      <button
-                        type="button"
-                        onClick={() => fileInputRef.current?.click()}
-                        className="text-primary hover:text-primary/80 underline font-medium transition-colors duration-200"
-                      >
-                        browse
-                      </button>
-                      {' '}or{' '}
-                      <button
-                        type="button"
-                        onClick={() => {
-                          setShowPasteArea(true);
-                          setTimeout(() => pasteTextareaRef.current?.focus(), 10);
-                        }}
-                        className="text-primary hover:text-primary/80 underline font-medium transition-colors duration-200"
-                      >
-                        paste code
-                      </button>
-                    </p>
-                    <p className="text-xs text-muted-foreground">
-                      Press Ctrl+V to paste files or code
-                    </p>
-                  </>
-                )}
-              </div>
-              <input
-                ref={fileInputRef}
-                type="file"
-                multiple
-                onChange={handleFileChange}
-                className="hidden"
-                accept="*/*"
-              />
-
-              {/* Paste Code Area - Only show when user clicks paste or presses Ctrl+V */}
-              {showPasteArea && (
-                <div className="space-y-2 mt-4 border-t border-border/30 pt-4">
-                  <div className="flex justify-between items-center mb-2.5">
-                    <label className="text-sm text-foreground font-semibold">Paste Code</label>
-                    <button
-                      type="button"
-                      onClick={handleCancelPaste}
-                      className="text-xs text-muted-foreground hover:text-primary underline transition-colors duration-200"
-                    >
-                      Cancel
-                    </button>
-                  </div>
-                  <textarea
-                    ref={pasteTextareaRef}
-                    value={pastedContent}
-                    onChange={(e) => {
-                      const newContent = e.target.value;
-                      setPastedContent(newContent);
-                      // Auto-detect language as user types/pastes
-                      if (newContent.trim().length > 10) {
-                        const detectedLang = detectLanguage(newContent);
-                        setLanguage(detectedLang);
-                      }
-                    }}
-                    onPaste={handleTextareaPaste}
-                    placeholder="Paste code here (Ctrl+V)..."
-                    rows={4}
-                    className="w-full px-4 py-3 border border-border/50 bg-white/80 text-foreground font-mono text-xs rounded-lg focus:outline-none focus:ring-2 focus:ring-primary/50 focus:border-primary shadow-modern transition-all duration-300 backdrop-blur-sm"
-                  />
-                  {pastedContent && (
-                    <div className="flex gap-2">
-                      <input
-                        type="text"
-                        value={pastedFileName}
-                        onChange={(e) => setPastedFileName(e.target.value)}
-                        placeholder="File name (optional)"
-                        className="flex-1 px-3 py-2 border border-border/50 bg-white/80 text-foreground text-sm rounded-lg focus:outline-none focus:ring-2 focus:ring-primary/50 focus:border-primary transition-all duration-300"
-                      />
-                      <select
-                        value={language}
-                        onChange={(e) => setLanguage(e.target.value)}
-                        className="px-3 py-2 border border-border/50 bg-white/80 text-foreground text-sm rounded-lg focus:outline-none focus:ring-2 focus:ring-primary/50 focus:border-primary transition-all duration-300"
-                      >
-                        <option value="text">text</option>
-                        {Object.entries(detectedLanguages)
-                          .filter(([ext, lang]) => lang !== "text") // Remove duplicate "text"
-                          .map(([ext, lang]) => (
-                            <option key={ext} value={lang}>
-                              {lang}
-                            </option>
-                          ))}
-                      </select>
-                      <button
-                        type="button"
-                        onClick={() => {
-                          handleAddPastedCode();
-                          // Keep paste area open for next paste
-                          setPastedContent("");
-                          setPastedFileName("");
-                          pasteTextareaRef.current?.focus();
-                        }}
-                        className="px-4 py-2 gradient-primary text-primary-foreground font-semibold hover:gradient-primary-hover rounded-lg transition-all duration-300 shadow-primary hover:shadow-primary-lg text-sm"
-                      >
-                        Add
-                      </button>
-                    </div>
-                  )}
-                </div>
-              )}
-            </div>
-          </div>
-
-          {/* File List */}
-          {(files.length > 0 || pastedCodeFiles.length > 0) && (
-            <div className="space-y-2 max-h-40 overflow-y-auto">
-              {pastedCodeFiles.map((file, index) => (
-                <div
-                  key={`pasted-${index}`}
-                  className="flex items-center justify-between p-2.5 border border-border/50 bg-white/80 rounded-lg backdrop-blur-sm"
-                >
-                  <span className="text-xs text-foreground truncate flex-1">📄 {file.filename}</span>
-                  <button
-                    type="button"
-                    onClick={() => removePastedCodeFile(index)}
-                    className="ml-2 text-muted-foreground hover:text-destructive text-sm transition-colors duration-200"
-                  >
-                    ×
-                  </button>
-                </div>
-              ))}
-              {files.map((file, index) => {
-                const ext = "." + file.name.split(".").pop()?.toLowerCase();
-                const isCodeFile = codeExtensions.includes(ext);
-                return (
-                  <div
-                    key={index}
-                    className="flex items-center justify-between p-2.5 border border-border/50 bg-white/80 rounded-lg backdrop-blur-sm"
-                  >
-                    <span className="text-xs text-foreground truncate flex-1">
-                      {isCodeFile ? '📄' : '📎'} {file.name}
-                    </span>
-                    <button
-                      type="button"
-                      onClick={() => removeFile(index)}
-                      className="ml-2 text-muted-foreground hover:text-destructive text-sm transition-colors duration-200"
-                    >
-                      ×
-                    </button>
-                  </div>
-                );
-              })}
-            </div>
-          )}
-
-          {error && (
-            <div className="animate-slide-up">
-              <p className="text-destructive text-sm font-medium">{error}</p>
-              {error.toLowerCase().includes('mime type') && error.toLowerCase().includes('not supported') && (
-                <p className="text-xs text-muted-foreground mt-2">
-                  Think this should be fixed? <a href="https://github.com/TryOmar/LabShare/issues/new" target="_blank" rel="noopener noreferrer" className="text-primary hover:text-primary/80 underline transition-colors duration-200">Open an issue</a> or <a href="https://forms.gle/25mEvcjTPrhA6THf9" target="_blank" rel="noopener noreferrer" className="text-primary hover:text-primary/80 underline transition-colors duration-200">report it</a>.
-                </p>
-              )}
-            </div>
-          )}
-
-          {/* Buttons */}
-          <div className="flex flex-col sm:flex-row gap-2 sm:gap-3 pt-3">
-            <button
-              type="submit"
-              disabled={loading || (files.length === 0 && pastedCodeFiles.length === 0)}
-              className="flex-1 gradient-primary text-primary-foreground font-semibold py-3 sm:py-3.5 text-sm sm:text-base rounded-lg hover:gradient-primary-hover hover:scale-[1.02] active:scale-[0.98] disabled:opacity-50 disabled:cursor-not-allowed transition-all duration-300 shadow-primary hover:shadow-primary-lg"
-            >
-              {loading ? "Uploading..." : "Upload"}
-            </button>
-            <button
-              type="button"
-              onClick={() => onClose(true)}
-              className="flex-1 border border-border/50 text-foreground font-semibold py-3 sm:py-3.5 text-sm sm:text-base rounded-lg hover:bg-accent/50 hover:border-primary/40 hover:text-primary transition-all duration-300 shadow-modern hover:shadow-primary/10 backdrop-blur-sm"
-            >
-              Cancel
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
-    </>
   );
 }
